Rename module 'action.decorators' to 'action.control'
(also remove one or two todo comments)
