+++ lib/plugins/default_subcmixer/__init__.py
@@ -23,7 +23,7 @@ import os
 import aiofiles
 import plugins
 import ujson as json
-from action.decorators import http_action
+from action.control import http_action
 from action.krequest import KRequest
 from action.model.corpus import CorpusActionModel
 from action.response import KResponse
