+++ lib/plugins/default_taghelper/__init__.py
@@ -39,7 +39,7 @@ element taghelper {
 from typing import Any, Dict
 
 import plugins
-from action.decorators import http_action
+from action.control import http_action
 from action.errors import UserReadableException
 from action.krequest import KRequest
 from action.model.corpus import CorpusActionModel
