+++ lib/plugins/default_user_items/__init__.py
@@ -17,7 +17,7 @@ import logging
 import l10n
 import plugins
 import ujson as json
-from action.decorators import http_action
+from action.control import http_action
 from action.krequest import KRequest
 from action.model.corpus import CorpusActionModel
 from action.response import KResponse
