+++ lib/plugins/masm_live_attributes/__init__.py
@@ -20,7 +20,7 @@
 import aiohttp
 import plugins
 import ujson
-from action.decorators import http_action
+from action.control import http_action
 from action.krequest import KRequest
 from action.model.corpus import CorpusActionModel
 from action.response import KResponse
