+++ lib/plugins/tree_corparch/__init__.py
@@ -58,7 +58,7 @@ from dataclasses import dataclass, field
 from typing import List, Union
 
 import plugins
-from action.decorators import http_action
+from action.control import http_action
 from action.krequest import KRequest
 from action.model.user import UserActionModel
 from action.plugin.ctx import AbstractCorpusPluginCtx, PluginCtx
