+++ lib/views/dispersion.py
@@ -17,7 +17,7 @@ from typing import List
 
 from action.argmapping.analytics import (
     CollFormArgs, CTFreqFormArgs, FreqFormArgs)
-from action.decorators import http_action
+from action.control import http_action
 from action.errors import ImmediateRedirectException, UserReadableException
 from action.krequest import KRequest
 from action.model.concordance import ConcActionModel
